Added explicit cast to DEB_ALLOC for the size (to size_t) to avoid warnings..
We might hide them though.. Maybe need to remove this...

+++ source/libdeb/include/debian/libdeb/mem.h
@@ -5,7 +5,7 @@
 extern "C" {
 #endif
 
-#define DEB_ALLOC(type, num) (##type) calloc(1, num)
+#define DEB_ALLOC(type, num) (##type) calloc(1, (size_t) num)
 
 #ifdef __cplusplus
 }
